MFC r311012 Remove some uses of the GCC __nonnull() attribute.
While the checks are considered useful, the attribute does dangerous
optimizations, removing NULL checks where they can be needed. Remove the
uses of this attribute introduced in r281130: the changes were inspired on
Google's bionic where this attribute is not used anymore.

In general, the __nonnull() attribute has fulfilled its roll and will be
replaced with the Clang _Nonnull qualifier in the future.


git-svn-id: https://svn.freebsd.org/base/stable/11@311234 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f