MFC r310817
sfxge(4): remove obsolete Wake-On-LAN support

Wake-on-lan is not supported in production on any of our adapters, as
they don't have the required AUX power connector. (It's possible that
AUX power is supplied to some of our ALOM or mezz adapters, but if so
then we've never implemented or tested WoL support.)

Sponsored by:   Solarflare Communications, Inc.


git-svn-id: https://svn.freebsd.org/base/stable/11@311490 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f