MFC r310655:
Fix bsnmpd sending/receiving with multi-homed configurations or INADDR_ANY used
as the listening address in snmpd_input(..)

Stash the IPv4 address of the receiver via the recv(..) callback and use it in
the send(..) callback for the transport by specifying IP_SENDSRCADDR for the
control message type.

Add sendmsg logic to the UDP transport's send(..) callback and use the
respective send(..) callback for the transport instead of calling sendto in
snmpd_input(..).

Obtained from:  Isilon OneFS (^/onefs/branches/BR_8_0_0_DEV@r507595)


git-svn-id: https://svn.freebsd.org/base/stable/10@312265 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f