MFC 307332,312086: Drop support for using mmap() with /dev/kmem.
307332:
Drop support for using mmap() with /dev/kmem.

Using the device pager with /dev/kmem is not stable since KVA mappings
are transient, but the device pager caches the PA associated with a
given offset forever.  Interestingly, mips' implementation of
memmap() already refused requests for /dev/kmem.

Note that kvm_read/kvm_write do not use mmap, but use read and write on
/dev/kmem, so this should not affect libkvm users.

312086:
Trim a few comments on platforms that did not implement mmap of /dev/kmem.

After r307332, no platforms implement mmap for /dev/kmem, so the lack of
it for these platforms is no longer unique.


git-svn-id: https://svn.freebsd.org/base/stable/11@312394 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f