MFC r311963: Remove writability requirement for single-mbuf, contiguous- range m_pulldown()
m_pulldown() only needs to determine if a mbuf is writable if it is going to
copy data into the data region of an existing mbuf. It does this to create a
contiguous data region in a single mbuf from multiple mbufs in the chain. If
the requested memory region is already contiguous and nothing needs to
change, the mbuf does not need to be writeable.


git-svn-id: https://svn.freebsd.org/base/stable/10@312442 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f