Configure the timer capture pin to input mode in the timer control register, in addition to configuring it as input with the pinmux driver.
There was a control register bit commented as "no desc in datasheet".  A
later revision of the manual reveals the bit to be an input/output control
for the timer pin.  In addition to configuring capture or pulse mode, you
apparently have to separately configure the pin direction in the timer
control register.

Before this change, the timer block was apparently driving a signal onto a
pad configured by pinmux as input.  Capture mode still accidentally worked
for me during testing because I was using a very strong signal source that
just out-muscled the weaker drive from the misconfigured pin.


git-svn-id: https://svn.freebsd.org/base/head@312859 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f